requirements.txt
about-time==4.2.1
absl-py==1.4.0
addict==2.4.0
alive-progress==3.0.1
astroid==2.12.13
asttokens==2.4.1
attrs==24.2.0
autograd==1.5
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
Brian2==2.5.1
certifi==2022.12.7
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
cma==3.2.2
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.0.6
control==0.9.4
csaps==1.1.0
cycler==0.11.0
Cython==0.29.32
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.23
docopt==0.6.2
et-xmlfile==1.1.0
etils==1.9.4
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.13.1
Flask==3.0.3
fonttools==4.38.0
fsspec==2024.9.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
glfw==2.5.7
grapheme==0.6.0
h5py==3.9.0
harold==1.0.3
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.3
ipywidgets==8.1.5
isort==5.11.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.4
labmaze==1.0.6
lazy-object-proxy==1.8.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
mpmath==1.2.1
mujoco==3.2.3
multidim_indexing==1.0.1
multiprocess==0.70.14
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nptyping==2.5.0
numexpr==2.8.5
numpy==1.24.3
open3d==0.18.0
openpyxl==3.1.2
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.5.0
platformdirs==2.6.0
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==4.22.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycollada==0.7.2
Pygments==2.18.0
pylint==2.15.9
pymoo==0.6.0.1
PyOpenGL==3.1.6
pyparsing==3.0.9
pyquaternion==0.9.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavefront==1.3.3
PyYAML==6.0
pyzmq==26.2.0
referencing==0.35.1
requests==2.28.2
retrying==1.3.4
rpds-py==0.20.0
Rtree==1.3.0
SALib==1.4.7
scikit-fmm==2024.9.16
scikit-learn==1.5.2
scipy==1.9.3
seaborn==0.13.0
simple-pid==1.0.1
six==1.16.0
smmap==5.0.0
soupsieve==2.6
stack-data==0.6.3
StrEnum==0.4.15
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.13.1
sysidentpy==0.3.1
tabulate==0.8.10
tenacity==9.0.0
threadpoolctl==3.1.0
tinycss2==1.4.0
tomli==2.0.1
tomlkit==0.11.6
tornado==6.4.1
tqdm==4.65.0
traitlets==5.14.3
trimesh==3.21.2
triton==3.1.0
typing_extensions==4.12.2
urllib3==1.26.15
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.14.1
yarg==0.1.9
zipp==3.20.2
